refactor(arch): 修复 core↔utils、utils↔services 分层违规
1. core/security.py 删除对 utils/notifier 的反向依赖，登录失败 IP 锁定通知职责上移到 api/auth.py。

2. core/database/inbox.py 不再反向依赖 utils；纯函数 notification_group_key 下沉到 inbox.py，utils/inbox_notifications.py 改为从 core 导入。

3. utils/data_util.py 删除对 services 子包 extractor 的反向 re-export，调用方直接 import services 子包。
